refactor provider clients into domain modules
