+++ src/main/gitlab/merge-request-lookup.ts
@@ -1,0 +1,183 @@
+import type { MRInfo } from '../../shared/gitlab-types'
+import { derivePipelineStatus, mapMRInfo } from './mappers'
+import {
+  acquire,
+  getGlabKnownHosts,
+  getProjectRef,
+  glabHostnameArgs,
+  glabRepoExecOptions,
+  glabExecFileAsync,
+  release,
+  type ProjectRef
+} from './gl-utils'
+import { encodedProject } from './project-path-encoding'
+import {
+  hasHostedReviewLocalGitOptions,
+  getHostedReviewLocalGitOptions,
+  type HostedReviewExecutionOptions
+} from '../source-control/hosted-review-git-options'
+import { shouldHideNonOpenReviewOnDefaultBranch } from '../source-control/repo-default-branch'
+
+type HostedReviewLocalGitOptions = ReturnType<typeof getHostedReviewLocalGitOptions>
+
+function hostedReviewLocalGitOptionArgs(
+  options: HostedReviewExecutionOptions = {}
+): [] | [HostedReviewLocalGitOptions] {
+  return hasHostedReviewLocalGitOptions(options) ? [getHostedReviewLocalGitOptions(options)] : []
+}
+
+export async function getProjectSlug(
+  repoPath: string,
+  connectionId?: string | null,
+  options: HostedReviewExecutionOptions = {}
+): Promise<ProjectRef | null> {
+  const localGitArgs = hostedReviewLocalGitOptionArgs(options)
+  const knownHosts = await getGlabKnownHosts(connectionId, localGitArgs[0])
+  return getProjectRef(repoPath, knownHosts, connectionId, ...localGitArgs)
+}
+
+/**
+ * Fetch a single merge request with pipeline status rolled up.
+ * Returns null when the MR doesn't exist or glab fails.
+ */
+export async function getMergeRequest(
+  repoPath: string,
+  iid: number,
+  connectionId?: string | null,
+  options: HostedReviewExecutionOptions = {}
+): Promise<MRInfo | null> {
+  const localGitArgs = hostedReviewLocalGitOptionArgs(options)
+  const localGitOptions = localGitArgs[0] ?? {}
+  const knownHosts = await getGlabKnownHosts(connectionId, localGitOptions)
+  const projectRef = await getProjectRef(repoPath, knownHosts, connectionId, ...localGitArgs)
+  await acquire()
+  try {
+    const args = projectRef
+      ? [
+          'api',
+          ...glabHostnameArgs(projectRef, connectionId),
+          `projects/${encodedProject(projectRef.path)}/merge_requests/${iid}`
+        ]
+      : ['mr', 'view', String(iid), '--output', 'json']
+    const { stdout } = await glabExecFileAsync(
+      args,
+      glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+    )
+    const data = JSON.parse(stdout) as Parameters<typeof mapMRInfo>[0] & {
+      head_pipeline?: { status?: string } | null
+      pipeline?: { status?: string } | null
+    }
+    // Why: older GitLab instances expose `pipeline` instead of `head_pipeline`; try both.
+    const pipelineStatus = derivePipelineStatus(data.head_pipeline ?? data.pipeline ?? null)
+    return mapMRInfo(data, pipelineStatus)
+  } catch {
+    return null
+  } finally {
+    release()
+  }
+}
+
+/**
+ * Find the merge request whose source branch matches the given name.
+ * Returns the most recently updated MR for the branch, or null when none exists.
+ */
+export async function getMergeRequestForBranch(
+  repoPath: string,
+  branch: string,
+  linkedMRIid?: number | null,
+  connectionId?: string | null,
+  options: HostedReviewExecutionOptions = {},
+  // Why: when true, a failed lookup throws instead of returning null, so callers never report a false not_found.
+  throwOnFailure = false
+): Promise<MRInfo | null> {
+  const branchName = branch.replace(/^refs\/heads\//, '')
+  if (!branchName && linkedMRIid == null) {
+    return null
+  }
+  const localGitArgs = hostedReviewLocalGitOptionArgs(options)
+  const localGitOptions = localGitArgs[0] ?? {}
+  const knownHosts = await getGlabKnownHosts(connectionId, localGitOptions)
+  const projectRef = await getProjectRef(repoPath, knownHosts, connectionId, ...localGitArgs)
+  if (!projectRef) {
+    return null
+  }
+  await acquire()
+  try {
+    if (branchName) {
+      const { stdout } = await glabExecFileAsync(
+        [
+          'api',
+          ...glabHostnameArgs(projectRef, connectionId),
+          // Why: GitLab does not proactively recompute merge status on list endpoints, so this row
+          // can sit at `unchecked` forever — and the sidebar merge button gates on MERGEABLE. Ask
+          // for the async recalculation (best-effort; ignored for non-Developers when
+          // `restrict_merge_status_recheck` is on) so polling converges instead of stalling.
+          `projects/${encodedProject(projectRef.path)}/merge_requests?source_branch=${encodeURIComponent(branchName)}&order_by=updated_at&sort=desc&per_page=1&with_merge_status_recheck=true`
+        ],
+        glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+      )
+      const data = JSON.parse(stdout) as (Parameters<typeof mapMRInfo>[0] & {
+        head_pipeline?: { status?: string } | null
+        pipeline?: { status?: string } | null
+      })[]
+      if (Array.isArray(data) && data.length > 0) {
+        const raw = data[0]
+        // Why: older GitLab list payloads expose `pipeline` instead of `head_pipeline`.
+        const pipelineStatus = derivePipelineStatus(raw.head_pipeline ?? raw.pipeline ?? null)
+        const info = mapMRInfo(raw, pipelineStatus)
+        // Why (#9171): discard a non-open implicit branch match on the repo
+        // default branch and fall through to the linked-iid fallback below.
+        const hideOnDefaultBranch = await shouldHideNonOpenReviewOnDefaultBranch({
+          state: info.state,
+          reviewNumber: info.number,
+          linkedReviewNumber: linkedMRIid,
+          branchName,
+          repoPath,
+          connectionId,
+          localGitOptions
+        })
+        if (!hideOnDefaultBranch) {
+          return info
+        }
+      }
+    }
+    if (typeof linkedMRIid !== 'number') {
+      return null
+    }
+    // Why: create-from-MR worktrees may rename the branch; fall back to the durable linked iid.
+    const { stdout } = await glabExecFileAsync(
+      [
+        'api',
+        ...glabHostnameArgs(projectRef, connectionId),
+        `projects/${encodedProject(projectRef.path)}/merge_requests/${linkedMRIid}`
+      ],
+      glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+    )
+    const raw = JSON.parse(stdout) as Parameters<typeof mapMRInfo>[0] & {
+      head_pipeline?: { status?: string } | null
+      pipeline?: { status?: string } | null
+    }
+    const pipelineStatus = derivePipelineStatus(raw.head_pipeline ?? raw.pipeline ?? null)
+    return mapMRInfo(raw, pipelineStatus)
+  } catch (error) {
+    if (throwOnFailure) {
+      throw error
+    }
+    return null
+  } finally {
+    release()
+  }
+}
+
+/**
+ * Like getMergeRequestForBranch but throws glab failures instead of returning null, so callers report 'unavailable' not a false "not found".
+ */
+export function getMergeRequestForBranchOrThrow(
+  repoPath: string,
+  branch: string,
+  linkedMRIid?: number | null,
+  connectionId?: string | null,
+  options: HostedReviewExecutionOptions = {}
+): Promise<MRInfo | null> {
+  return getMergeRequestForBranch(repoPath, branch, linkedMRIid, connectionId, options, true)
+}
