+++ src/main/gitlab/merge-request-project-resolution.ts
@@ -1,0 +1,33 @@
+import type { IssueSourcePreference } from '../../shared/repo-types'
+import {
+  getGlabKnownHosts,
+  resolveIssueSource,
+  type LocalGitExecOptions,
+  type ProjectRef
+} from './gl-utils'
+
+export async function withProjectRef<T>(
+  repoPath: string,
+  preference: IssueSourcePreference | undefined,
+  connectionId: string | null | undefined,
+  explicitProjectRef: ProjectRef | null | undefined,
+  fn: (projectRef: ProjectRef, repoFlag: string) => Promise<T>,
+  fallback: T,
+  localGitOptions: LocalGitExecOptions = {}
+): Promise<T> {
+  const projectRef =
+    explicitProjectRef ??
+    (
+      await resolveIssueSource(
+        repoPath,
+        preference,
+        await getGlabKnownHosts(connectionId, localGitOptions),
+        connectionId,
+        localGitOptions
+      )
+    ).source
+  if (!projectRef) {
+    return fallback
+  }
+  return fn(projectRef, projectRef.path)
+}
