+++ src/main/gitlab/merge-request-state-mutations.ts
@@ -1,0 +1,143 @@
+import type { IssueSourcePreference } from '../../shared/repo-types'
+import {
+  acquire,
+  glabHostnameArgs,
+  glabRepoExecOptions,
+  glabExecFileAsync,
+  release,
+  type LocalGitExecOptions,
+  type ProjectRef
+} from './gl-utils'
+import { withProjectRef } from './merge-request-project-resolution'
+
+export async function closeMR(
+  repoPath: string,
+  iid: number,
+  preference?: IssueSourcePreference,
+  connectionId?: string | null,
+  projectRef?: ProjectRef | null,
+  localGitOptions: LocalGitExecOptions = {}
+): Promise<{ ok: true } | { ok: false; error: string }> {
+  return withProjectRef<{ ok: true } | { ok: false; error: string }>(
+    repoPath,
+    preference,
+    connectionId,
+    projectRef,
+    async (projectRef, repoFlag) => {
+      await acquire()
+      try {
+        await glabExecFileAsync(
+          [
+            'mr',
+            'close',
+            String(iid),
+            '-R',
+            repoFlag,
+            ...glabHostnameArgs(projectRef, connectionId)
+          ],
+          glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+        )
+        return { ok: true }
+      } catch (err) {
+        const msg = err instanceof Error ? err.message : String(err)
+        // Why: glab exits non-zero when the MR is already closed — treat as success.
+        if (msg.toLowerCase().includes('already')) {
+          return { ok: true }
+        }
+        return { ok: false, error: msg }
+      } finally {
+        release()
+      }
+    },
+    { ok: false, error: 'Could not resolve GitLab project for this repository' },
+    localGitOptions
+  )
+}
+
+export async function reopenMR(
+  repoPath: string,
+  iid: number,
+  preference?: IssueSourcePreference,
+  connectionId?: string | null,
+  projectRef?: ProjectRef | null,
+  localGitOptions: LocalGitExecOptions = {}
+): Promise<{ ok: true } | { ok: false; error: string }> {
+  return withProjectRef<{ ok: true } | { ok: false; error: string }>(
+    repoPath,
+    preference,
+    connectionId,
+    projectRef,
+    async (projectRef, repoFlag) => {
+      await acquire()
+      try {
+        await glabExecFileAsync(
+          [
+            'mr',
+            'reopen',
+            String(iid),
+            '-R',
+            repoFlag,
+            ...glabHostnameArgs(projectRef, connectionId)
+          ],
+          glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+        )
+        return { ok: true }
+      } catch (err) {
+        const msg = err instanceof Error ? err.message : String(err)
+        if (msg.toLowerCase().includes('already')) {
+          return { ok: true }
+        }
+        return { ok: false, error: msg }
+      } finally {
+        release()
+      }
+    },
+    { ok: false, error: 'Could not resolve GitLab project for this repository' },
+    localGitOptions
+  )
+}
+
+export async function mergeMR(
+  repoPath: string,
+  iid: number,
+  method: 'merge' | 'squash' | 'rebase' = 'merge',
+  preference?: IssueSourcePreference,
+  connectionId?: string | null,
+  projectRef?: ProjectRef | null,
+  localGitOptions: LocalGitExecOptions = {}
+): Promise<{ ok: true } | { ok: false; error: string }> {
+  return withProjectRef<{ ok: true } | { ok: false; error: string }>(
+    repoPath,
+    preference,
+    connectionId,
+    projectRef,
+    async (projectRef, repoFlag) => {
+      await acquire()
+      try {
+        // Why: omitting both --squash and --rebase yields a regular merge commit.
+        const methodFlag =
+          method === 'squash' ? ['--squash'] : method === 'rebase' ? ['--rebase'] : []
+        await glabExecFileAsync(
+          [
+            'mr',
+            'merge',
+            String(iid),
+            '-R',
+            repoFlag,
+            '--yes',
+            ...methodFlag,
+            ...glabHostnameArgs(projectRef, connectionId)
+          ],
+          glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+        )
+        return { ok: true }
+      } catch (err) {
+        return { ok: false, error: err instanceof Error ? err.message : String(err) }
+      } finally {
+        release()
+      }
+    },
+    { ok: false, error: 'Could not resolve GitLab project for this repository' },
+    localGitOptions
+  )
+}
