+++ src/main/gitlab/pipeline-job-mutations.ts
@@ -1,0 +1,128 @@
+import type {
+  GitLabJobTraceResult,
+  GitLabPipelineJob,
+  GitLabRetryJobResult
+} from '../../shared/gitlab-types'
+import type { IssueSourcePreference } from '../../shared/repo-types'
+import {
+  acquire,
+  classifyGlabError,
+  classifyJobLogError,
+  glabHostnameArgs,
+  glabRepoExecOptions,
+  glabExecFileAsync,
+  isMissingJobLogError,
+  release,
+  type LocalGitExecOptions,
+  type ProjectRef
+} from './gl-utils'
+import { encodedProject } from './project-path-encoding'
+import { withProjectRef } from './merge-request-project-resolution'
+
+function mapRetriedPipelineJob(
+  data: {
+    id?: number
+    pipeline?: { id?: number | null } | null
+    name?: string
+    stage?: string
+    status?: string
+    web_url?: string
+    duration?: number | null
+  },
+  fallbackJobId: number
+): GitLabPipelineJob {
+  return {
+    id: data.id ?? fallbackJobId,
+    ...(typeof data.pipeline?.id === 'number' ? { pipelineId: data.pipeline.id } : {}),
+    name: data.name ?? '',
+    stage: data.stage ?? '',
+    status: data.status ?? '',
+    webUrl: data.web_url ?? '',
+    duration: typeof data.duration === 'number' ? data.duration : null
+  }
+}
+
+export async function getJobTrace(
+  repoPath: string,
+  jobId: number,
+  preference?: IssueSourcePreference,
+  connectionId?: string | null,
+  projectRef?: ProjectRef | null,
+  localGitOptions: LocalGitExecOptions = {}
+): Promise<GitLabJobTraceResult> {
+  return withProjectRef<GitLabJobTraceResult>(
+    repoPath,
+    preference,
+    connectionId,
+    projectRef,
+    async (projectRef) => {
+      await acquire()
+      try {
+        const { stdout } = await glabExecFileAsync(
+          [
+            'api',
+            ...glabHostnameArgs(projectRef, connectionId),
+            `projects/${encodedProject(projectRef.path)}/jobs/${jobId}/trace`
+          ],
+          glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+        )
+        return { ok: true, trace: stdout }
+      } catch (err) {
+        const msg = err instanceof Error ? err.message : String(err)
+        // A job with no log is an empty log, not a failure: surfacing the 404 would
+        // pin an error on the Checks row for a job canceled before it started.
+        if (isMissingJobLogError(msg)) {
+          return { ok: true, trace: '' }
+        }
+        return { ok: false, error: classifyJobLogError(msg).message }
+      } finally {
+        release()
+      }
+    },
+    { ok: false, error: 'Could not resolve GitLab project for this repository' },
+    localGitOptions
+  )
+}
+
+export async function retryJob(
+  repoPath: string,
+  jobId: number,
+  preference?: IssueSourcePreference,
+  connectionId?: string | null,
+  projectRef?: ProjectRef | null,
+  localGitOptions: LocalGitExecOptions = {}
+): Promise<GitLabRetryJobResult> {
+  return withProjectRef<GitLabRetryJobResult>(
+    repoPath,
+    preference,
+    connectionId,
+    projectRef,
+    async (projectRef) => {
+      await acquire()
+      try {
+        const { stdout } = await glabExecFileAsync(
+          [
+            'api',
+            ...glabHostnameArgs(projectRef, connectionId),
+            '-X',
+            'POST',
+            `projects/${encodedProject(projectRef.path)}/jobs/${jobId}/retry`
+          ],
+          glabRepoExecOptions(repoPath, connectionId, localGitOptions)
+        )
+        const trimmed = stdout.trim()
+        return {
+          ok: true,
+          ...(trimmed ? { job: mapRetriedPipelineJob(JSON.parse(trimmed), jobId) } : {})
+        }
+      } catch (err) {
+        const msg = err instanceof Error ? err.message : String(err)
+        return { ok: false, error: classifyGlabError(msg).message }
+      } finally {
+        release()
+      }
+    },
+    { ok: false, error: 'Could not resolve GitLab project for this repository' },
+    localGitOptions
+  )
+}
