+++ src/main/jira/jira-issue-create-metadata.ts
@@ -1,0 +1,160 @@
+import type {
+  JiraCreateField,
+  JiraIssueType,
+  JiraPriority,
+  JiraUser
+} from '../../shared/jira-types'
+import { acquire, release } from './request-queue'
+import { apiBasePath, jiraRequest } from './authenticated-request'
+import { clearToken, getClients, isAuthError } from './client'
+import {
+  getCreateFieldRecords,
+  mapCreateField,
+  mapIssueType,
+  mapPriority,
+  mapUser
+} from './jira-issue-mapping'
+import {
+  asFiniteNumber,
+  fetchPagedRecords,
+  shouldFetchNextPage,
+  type JiraPagedResponse,
+  type JiraRecord
+} from './jira-record-pages'
+
+export async function listIssueTypes(
+  projectIdOrKey: string,
+  siteId?: string | null
+): Promise<JiraIssueType[]> {
+  const entry = getClients(siteId)[0]
+  if (!entry) {
+    return []
+  }
+  await acquire()
+  try {
+    const issueTypes = await fetchPagedRecords(entry, 'issueTypes', (startAt, maxResults) => {
+      const params = new URLSearchParams({
+        maxResults: String(maxResults),
+        startAt: String(startAt)
+      })
+      // Per-project createmeta paths exist on Server/DC from Jira 8.4 onward.
+      return `${apiBasePath(entry.site)}/issue/createmeta/${encodeURIComponent(
+        projectIdOrKey
+      )}/issuetypes?${params.toString()}`
+    })
+    return issueTypes.map(mapIssueType)
+  } catch (error) {
+    if (isAuthError(error)) {
+      clearToken(entry.site.id)
+      throw error
+    }
+    console.warn('[jira] listIssueTypes failed:', error)
+    return []
+  } finally {
+    release()
+  }
+}
+
+export async function listCreateFields(
+  projectIdOrKey: string,
+  issueTypeId: string,
+  siteId?: string | null
+): Promise<JiraCreateField[]> {
+  const entry = getClients(siteId)[0]
+  if (!entry) {
+    return []
+  }
+  await acquire()
+  try {
+    const fields: JiraCreateField[] = []
+    let startAt = 0
+    const maxResults = 100
+    for (let guard = 0; guard < 100; guard += 1) {
+      const params = new URLSearchParams({
+        maxResults: String(maxResults),
+        startAt: String(startAt)
+      })
+      const response = await jiraRequest<JiraPagedResponse<JiraRecord>>(
+        entry,
+        `${apiBasePath(entry.site)}/issue/createmeta/${encodeURIComponent(
+          projectIdOrKey
+        )}/issuetypes/${encodeURIComponent(issueTypeId)}?${params.toString()}`
+      )
+      const records = getCreateFieldRecords(response)
+      fields.push(
+        ...records
+          .map((record) => mapCreateField(record))
+          .filter((field): field is JiraCreateField => field !== null)
+      )
+      if (!shouldFetchNextPage(response, startAt, records, maxResults)) {
+        break
+      }
+      startAt += asFiniteNumber(response.maxResults) ?? maxResults
+    }
+    return fields
+  } catch (error) {
+    if (isAuthError(error)) {
+      clearToken(entry.site.id)
+      throw error
+    }
+    console.warn('[jira] listCreateFields failed:', error)
+    return []
+  } finally {
+    release()
+  }
+}
+
+export async function listPriorities(siteId?: string | null): Promise<JiraPriority[]> {
+  const entry = getClients(siteId)[0]
+  if (!entry) {
+    return []
+  }
+  await acquire()
+  try {
+    const response = await jiraRequest<JiraRecord[]>(entry, `${apiBasePath(entry.site)}/priority`)
+    return response.map(mapPriority).filter((priority): priority is JiraPriority => !!priority)
+  } catch (error) {
+    if (isAuthError(error)) {
+      clearToken(entry.site.id)
+      throw error
+    }
+    console.warn('[jira] listPriorities failed:', error)
+    return []
+  } finally {
+    release()
+  }
+}
+
+export async function listAssignableUsers(
+  key: string,
+  query?: string,
+  siteId?: string | null
+): Promise<JiraUser[]> {
+  const entry = getClients(siteId)[0]
+  if (!entry) {
+    return []
+  }
+  const isServer = entry.site.authType === 'server'
+  const params = new URLSearchParams({ issueKey: key, maxResults: '50' })
+  if (query?.trim()) {
+    // Server/DC filters assignable users by `username`; `query` is Cloud-only.
+    params.set(isServer ? 'username' : 'query', query.trim())
+  }
+  await acquire()
+  try {
+    const response = await jiraRequest<JiraRecord[]>(
+      entry,
+      `${apiBasePath(entry.site)}/user/assignable/search?${params.toString()}`
+    )
+    return response.map(mapUser).filter((user): user is JiraUser => !!user)
+  } catch (error) {
+    if (isAuthError(error)) {
+      clearToken(entry.site.id)
+      throw error
+    }
+    console.warn('[jira] listAssignableUsers failed:', error)
+    return []
+  } finally {
+    release()
+  }
+}
