+++ src/main/jira/jira-issue-search.ts
@@ -1,0 +1,122 @@
+import type { JiraIssue, JiraIssueFilter, JiraSiteSelection } from '../../shared/jira-types'
+import { acquire, release } from './request-queue'
+import { apiBasePath, jiraRequest, type JiraClientForSite } from './authenticated-request'
+import { clearToken, getClients, isAuthError } from './client'
+import { ISSUE_LIST_FIELDS, mapJiraIssue } from './jira-issue-mapping'
+import type { JiraSearchResponse } from './jira-record-pages'
+import {
+  shouldSurfaceSiteFailure,
+  toIssueSearchFailureError,
+  withJiraDeadline,
+  type JiraIssueSearchFailure
+} from './jira-read-failure'
+
+const ISSUE_SEARCH_TIMEOUT_MS = 30_000
+
+function clampLimit(limit: number | undefined, fallback = 30): number {
+  return Math.min(Math.max(1, Number.isFinite(limit) ? Number(limit) : fallback), 100)
+}
+
+function sortAndLimitIssues(issues: JiraIssue[], limit: number): JiraIssue[] {
+  return issues
+    .sort((a, b) => new Date(b.updatedAt).getTime() - new Date(a.updatedAt).getTime())
+    .slice(0, limit)
+}
+
+function filterToJql(filter: JiraIssueFilter): string {
+  if (filter === 'assigned') {
+    return 'assignee = currentUser() AND resolution = Unresolved ORDER BY updated DESC'
+  }
+  if (filter === 'reported') {
+    return 'reporter = currentUser() AND resolution = Unresolved ORDER BY updated DESC'
+  }
+  if (filter === 'done') {
+    return 'assignee = currentUser() AND resolution IS NOT EMPTY ORDER BY updated DESC'
+  }
+  return 'resolution = Unresolved ORDER BY updated DESC'
+}
+
+async function searchIssuesForClient(
+  entry: JiraClientForSite,
+  jql: string,
+  limit: number,
+  signal?: AbortSignal
+): Promise<JiraIssue[]> {
+  // Server/DC only has the classic /search resource; /search/jql is Cloud-only.
+  const searchPath =
+    entry.site.authType === 'server'
+      ? `${apiBasePath(entry.site)}/search`
+      : '/rest/api/3/search/jql'
+  const result = await jiraRequest<JiraSearchResponse>(entry, searchPath, {
+    method: 'POST',
+    body: JSON.stringify({
+      jql,
+      maxResults: limit,
+      fields: ISSUE_LIST_FIELDS
+    }),
+    signal
+  })
+  return (result.issues ?? []).map((issue) => mapJiraIssue(entry.site, issue))
+}
+
+export async function listIssues(
+  filter: JiraIssueFilter = 'assigned',
+  limit = 30,
+  siteId?: JiraSiteSelection | null
+): Promise<JiraIssue[]> {
+  return searchIssues(filterToJql(filter), limit, siteId)
+}
+
+export async function searchIssues(
+  jql: string,
+  limit = 30,
+  siteId?: JiraSiteSelection | null,
+  signal?: AbortSignal
+): Promise<JiraIssue[]> {
+  const entries = getClients(siteId)
+  if (entries.length === 0 || !jql.trim()) {
+    return []
+  }
+  const safeLimit = clampLimit(limit)
+  const failures: (JiraIssueSearchFailure | undefined)[] = Array.from({ length: entries.length })
+  const surfaceSiteFailure = shouldSurfaceSiteFailure(siteId, entries.length)
+  const results = await withJiraDeadline(signal, ISSUE_SEARCH_TIMEOUT_MS, (requestSignal) =>
+    Promise.all(
+      entries.map(async (entry, index) => {
+        // Why: queueing on an abandoned search would keep occupying the shared Jira pool.
+        await acquire(requestSignal)
+        try {
+          return await searchIssuesForClient(entry, jql.trim(), safeLimit, requestSignal)
+        } catch (error) {
+          if (requestSignal.aborted) {
+            // Abandoned by the caller: not a site failure, so don't clear tokens or mask a real one.
+            throw error
+          }
+          const authFailure = isAuthError(error)
+          if (authFailure) {
+            clearToken(entry.site.id)
+          }
+          if (surfaceSiteFailure) {
+            throw toIssueSearchFailureError(error)
+          }
+          console.warn('[jira] searchIssues failed:', error)
+          failures[index] = { error: toIssueSearchFailureError(error), auth: authFailure }
+          return [] as JiraIssue[]
+        } finally {
+          release()
+        }
+      })
+    )
+  )
+  // 'all' fan-out: only surface an error when every connected site failed, so a
+  // partial success (or a genuinely empty result) is not reported as an error.
+  const recordedFailures = failures.filter(
+    (failure): failure is JiraIssueSearchFailure => failure !== undefined
+  )
+  if (recordedFailures.length === entries.length) {
+    throw (recordedFailures.find((failure) => !failure.auth) ?? recordedFailures[0]).error
+  }
+  return entries.length === 1
+    ? results.flat().slice(0, safeLimit)
+    : sortAndLimitIssues(results.flat(), safeLimit)
+}
