+++ src/main/linear/projects.test.ts
@@ -388,6 +388,8 @@ describe('Linear project queries', () => {
         targetDate: '2026-08-01'
       }
     })
+    expect(rawRequest.mock.calls[0]?.[0]).toContain('mutation OrcaLinearProjectCreate')
+    expect(rawRequest.mock.calls[0]?.[0]).toContain('projectCreate(input: $input)')
   })
 
   it('lets manual custom view list refresh bypass older in-flight reads', async () => {
